Allow Rubinius to fail on travis
The rspec call does not terminate on Rubinius even if all tests are
green and the test suite actually has completed. Some of the tests
introduced in this feature branch seem to reveal this error.

If only confirmations_controller_spec and repositories_controller_spec
are there, rspec does not terminate. If we remove the POST specs from
the confirmations_controller_spec, then rspec terminates. Strangely,
Rspec also terminates if we remove the usages of the `have_http_status`
matcher from the repositories_controller_spec instead.