127 setup devise confirmation and notification mails (#128)
* Update ontohub-models.

* Configure ActionMailer.

* Setup devise confirmable module.

* Update specs to confirmable.

* Fix seeds.

* Load environment.local.rb before initialization.

* Add data directory to gitignore.

* Setup devise recoverable module.

* Add custom mailer and views for devise.

* Fix registrations#update schema.

* Send email changed notifications.

* Add specs for mailers.

* Make sessions controller implementation consistent to other devise controllers.

* Setup devise lockable module.

* Obey linter.

* Add failing recoverable specs.

* Allow Rubinius to fail on travis

The rspec call does not terminate on Rubinius even if all tests are
green and the test suite actually has completed. Some of the tests
introduced in this feature branch seem to reveal this error.

If only confirmations_controller_spec and repositories_controller_spec
are there, rspec does not terminate. If we remove the POST specs from
the confirmations_controller_spec, then rspec terminates. Strangely,
Rspec also terminates if we remove the usages of the `have_http_status`
matcher from the repositories_controller_spec instead.

* Run the seeds in travis.

* Fix time zone problem of sequel.

* Fix GET to PATCH change.

* Fix typos.